refactor: extract turn pipeline completion and ratchet 250/300
Share one halt/full-run completion helper and spy-line sibling, densify two fat suites, and lower turn lib/test size gates.
