+++ packages/colonizethis_turn/test/turn/last_turn_intelligence_digest_test.dart
@@ -5,6 +5,7 @@ import 'package:colonizethis_turn/colonizethis_turn.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
 
 import '../support/turn_news_digest_test_support.dart';
+import 'last_turn_intelligence_digest_test_cases.dart';
 
 void main() {
   group('buildLastTurnIntelligenceDigest', () {
@@ -28,7 +29,7 @@ void main() {
             ),
             newWorld: const RegionData(),
           ),
-          players: _threeGps,
+          players: lastTurnIntelThreeGps,
           diplomacyRelations: const [
             DiplomacyRelation(
               factionId1: 'gp2',
@@ -83,173 +84,7 @@ void main() {
       },
     );
 
-    test(
-      'Given no spy in France When France unlocks tech and fights Then no spy block',
-      () {
-        final start = _franceSpainGame(turn: 2, spyInFrance: false);
-        final end = _franceSpainGame(
-          turn: 3,
-          spyInFrance: false,
-          franceTech: const {kTechIdCropRotation: true},
-        );
-        final news = const TurnNewsDigest(resolvedTurnNumber: 2, lines: []);
-        final digest = buildLastTurnIntelligenceDigest(
-          start: start,
-          end: end,
-          worldNews: news,
-          turnEvents: const [
-            CombatResultEvent(
-              provinceId: 'oldWorld|fr1',
-              attackerId: 'france',
-              defenderId: 'gp3',
-              outcomeName: 'attackerVictory',
-              winnerId: 'france',
-              turnNumber: 2,
-            ),
-          ],
-        );
-        expect(digest.spyReportsFor('gp1'), isEmpty);
-      },
-    );
-
-    test(
-      'Given spy remaining in France When France techs and declares war Then spy block',
-      () {
-        final start = _franceSpainGame(turn: 2, spyInFrance: true);
-        final end =
-            _franceSpainGame(
-              turn: 3,
-              spyInFrance: true,
-              franceTech: const {kTechIdCropRotation: true},
-            ).copyWith(
-              diplomaticHistoryEvents: const [
-                DiplomaticEvent(
-                  turn: 2,
-                  intraTurnIndex: 0,
-                  type: DiplomaticEventType.declareWar,
-                  participants: {'france', 'gp3'},
-                  fromFactionId: 'france',
-                  toFactionId: 'gp3',
-                ),
-              ],
-            );
-        final digest = buildLastTurnIntelligenceDigest(
-          start: start,
-          end: end,
-          worldNews: const TurnNewsDigest(resolvedTurnNumber: 2, lines: []),
-        );
-        final block = digest.spyReportsFor('gp1').single;
-        expect(block.courtFactionId, 'france');
-        expect(
-          block.lines.any(
-            (l) =>
-                l.kind == IntelligenceSpyKind.diplomatic &&
-                l.diplomaticType == DiplomaticEventType.declareWar &&
-                l.toFactionId == 'gp3',
-          ),
-          isTrue,
-        );
-        expect(
-          block.lines.any(
-            (l) =>
-                l.kind == IntelligenceSpyKind.researchComplete &&
-                l.techId == kTechIdCropRotation,
-          ),
-          isTrue,
-        );
-        expect(block.lines.any((l) => l.techId == 'hiddenAgenda'), isFalse);
-      },
-    );
-
-    test(
-      'Given last spy left France When digest builds Then France spy omitted',
-      () {
-        final start = _franceSpainGame(turn: 2, spyInFrance: true);
-        final end = _franceSpainGame(turn: 3, spyInFrance: false).copyWith(
-          diplomaticHistoryEvents: const [
-            DiplomaticEvent(
-              turn: 2,
-              intraTurnIndex: 0,
-              type: DiplomaticEventType.declareWar,
-              participants: {'france', 'gp3'},
-              fromFactionId: 'france',
-              toFactionId: 'gp3',
-            ),
-          ],
-        );
-        final digest = buildLastTurnIntelligenceDigest(
-          start: start,
-          end: end,
-          worldNews: const TurnNewsDigest(resolvedTurnNumber: 2, lines: []),
-        );
-        expect(digest.spyReportsFor('gp1'), isEmpty);
-      },
-    );
-
-    test(
-      'Given alliance formed When digest builds Then world lines include alliance',
-      () {
-        final start = _franceSpainGame(turn: 2, spyInFrance: false);
-        final end = _franceSpainGame(turn: 3, spyInFrance: false).copyWith(
-          diplomaticHistoryEvents: const [
-            DiplomaticEvent(
-              turn: 2,
-              intraTurnIndex: 0,
-              type: DiplomaticEventType.allianceFormed,
-              participants: {'france', 'gp3'},
-              fromFactionId: 'france',
-              toFactionId: 'gp3',
-            ),
-          ],
-        );
-        final digest = buildLastTurnIntelligenceDigest(
-          start: start,
-          end: end,
-          worldNews: const TurnNewsDigest(resolvedTurnNumber: 2, lines: []),
-        );
-        expect(
-          digest.worldLines.any(
-            (l) =>
-                l.kind == IntelligenceWorldKind.allianceFormed &&
-                l.factionIdA == 'france' &&
-                l.factionIdB == 'gp3',
-          ),
-          isTrue,
-        );
-      },
-    );
-
-    test('Given spy remaining When France fights Then spy combat line', () {
-      final start = _franceSpainGame(turn: 2, spyInFrance: true);
-      final end = _franceSpainGame(turn: 3, spyInFrance: true);
-      final digest = buildLastTurnIntelligenceDigest(
-        start: start,
-        end: end,
-        worldNews: const TurnNewsDigest(resolvedTurnNumber: 2, lines: []),
-        turnEvents: const [
-          CombatResultEvent(
-            provinceId: 'oldWorld|fr1',
-            attackerId: 'france',
-            defenderId: 'gp3',
-            outcomeName: 'attackerVictory',
-            winnerId: 'france',
-            turnNumber: 2,
-          ),
-        ],
-      );
-      expect(
-        digest
-            .spyReportsFor('gp1')
-            .single
-            .lines
-            .any(
-              (l) =>
-                  l.kind == IntelligenceSpyKind.combat &&
-                  l.provinceId == 'oldWorld|fr1',
-            ),
-        isTrue,
-      );
-    });
+    registerLastTurnIntelSpyCases();
 
     test('Given new digest When persist Then previous digest replaced', () {
       final start = turnNewsMinimalGame(turn: 2).copyWith(
@@ -286,64 +121,4 @@ void main() {
       expect(game.lastTurnIntelligenceDigest, isNull);
     });
   });
-}
-
-const _threeGps = [
-  Player(id: 'gp1', displayName: 'England', isHuman: true, treasury: 0),
-  Player(id: 'gp2', displayName: 'Spain', isHuman: false, treasury: 0),
-  Player(id: 'gp3', displayName: 'France', isHuman: false, treasury: 0),
-];
-
-Game _franceSpainGame({
-  required int turn,
-  required bool spyInFrance,
-  Map<String, bool> franceTech = const {},
-}) {
-  return Game(
-    id: 'g',
-    worldState: WorldState(
-      turnState: TurnState(phase: TurnPhase.orders, turnNumber: turn),
-      oldWorld: RegionData(
-        provinces: const [
-          Province(
-            id: 'oldWorld|fr1',
-            regionId: 'oldWorld',
-            ownerId: 'france',
-            displayName: 'Paris',
-          ),
-        ],
-        units: [
-          if (spyInFrance)
-            Unit(
-              id: 'spy1',
-              type: kUnitTypeSpy,
-              ownerId: 'gp1',
-              locationProvinceId: 'oldWorld|fr1',
-            ),
-        ],
-      ),
-      newWorld: const RegionData(),
-    ),
-    players: [
-      const Player(
-        id: 'gp1',
-        displayName: 'England',
-        isHuman: true,
-        treasury: 0,
-      ),
-      Player(
-        id: 'france',
-        displayName: 'France',
-        isHuman: false,
-        treasury: 0,
-        techUnlocked: franceTech,
-      ),
-      const Player(
-        id: 'gp3',
-        displayName: 'Spain',
-        isHuman: false,
-        treasury: 0,
-      ),
-    ],
-  );
 }
