+++ tool/ct_repo_lint_manifest.yaml
@@ -909,14 +909,14 @@ rules:
 
   - rule_id: repo.turn_lib_file_size
     group: structure
-    title: colonizethis_turn lib files must stay at or below the wave-7 300 physical-line ceiling (Refs #4113, #4342)
+    title: colonizethis_turn lib files must stay at or below the wave-8 250 physical-line ceiling (Refs #4113, #4342, #4583)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_turn_lib_file_size.dart
 
   - rule_id: repo.turn_test_file_size
     group: testing
-    title: colonizethis_turn non-support test files must stay at or below the wave-4 400 physical-line ceiling (Refs #4113)
+    title: colonizethis_turn non-support test files must stay at or below the wave-8 300 physical-line ceiling (Refs #4113, #4583)
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_turn_test_file_size.dart
